server/requirements.txt
pandas
fastapi
uvicorn
pydantic
pytest
psycopg2-binary
pydantic-settings
python-dotenv
types-psycopg2